+++ src/Command/IndexCommand.php
@@ -17,10 +17,11 @@ declare(strict_types=1);
 namespace DatabaseBackup\Command;
 
 use Cake\Console\Arguments;
+use Cake\Console\BaseCommand;
 use Cake\Console\ConsoleIo;
 use Cake\Console\ConsoleOptionParser;
 use Cake\I18n\Number;
-use DatabaseBackup\Console\Command;
+use DatabaseBackup\Compression;
 use DatabaseBackup\Utility\BackupManager;
 
 /**
@@ -28,7 +29,7 @@ use DatabaseBackup\Utility\BackupManager;
  *
  * @see https://github.com/mirko-pagliai/cakephp-database-backup/wiki/How-to-use-commands#index
  */
-class IndexCommand extends Command
+class IndexCommand extends BaseCommand
 {
     /**
      * @inheritDoc
@@ -52,8 +53,6 @@ class IndexCommand extends Command
      */
     public function execute(Arguments $args, ConsoleIo $io): void
     {
-        parent::execute($args, $io);
-
         $backups = BackupManager::index();
         $io->out(__d('database_backup', 'Backup files found: {0}', $backups->count()));
 
@@ -63,7 +62,6 @@ class IndexCommand extends Command
 
         $headers = [
             __d('database_backup', 'Filename'),
-            __d('database_backup', 'Extension'),
             __d('database_backup', 'Compression'),
             __d('database_backup', 'Size'),
             __d('database_backup', 'Datetime'),
@@ -71,7 +69,10 @@ class IndexCommand extends Command
 
         $rows = $backups
             ->map(fn (array $backup): array => array_merge($backup, [
-               'compression' => $backup['compression'] ?: '',
+               'compression' => match ($backup['compression']) {
+                    Compression::None => '',
+                    default => lcfirst($backup['compression']->name)
+               },
                'datetime' => $backup['datetime']->nice(),
                'size' => Number::toReadableSize($backup['size']),
             ]))