+++ src/Driver/AbstractDriver.php
@@ -20,6 +20,8 @@ use Cake\Core\Configure;
 use Cake\Event\EventDispatcherTrait;
 use Cake\Event\EventListenerInterface;
 use DatabaseBackup\BackupTrait;
+use DatabaseBackup\Compression;
+use DatabaseBackup\OperationType;
 use LogicException;
 
 /**
@@ -69,22 +71,14 @@ abstract class AbstractDriver implements EventListenerInterface
      * These executables are not yet final, use instead `getExportExecutable()` and `getImportExecutable()` methods to
      *  have the final executables, including compression.
      *
-     * @param string $type Type or the request operation (`export` or `import`)
+     * @param \DatabaseBackup\OperationType $OperationType
      * @return string
-     * @throws \LogicException
      */
-    private function getExecutable(string $type): string
+    private function getExecutable(OperationType $OperationType): string
     {
-        // @codeCoverageIgnoreStart
-        if (!in_array($type, ['export', 'import'])) {
-            throw new LogicException(
-                __d('database_backup', '`{0}` parameter should be `{1}` or `{2}`', '$type', 'export', 'import')
-            );
-        }
-        // @codeCoverageIgnoreEnd
         $driverName = strtolower($this->getDriverName());
         $replacements = [
-            '{{BINARY}}' => escapeshellarg($this->getBinary(DATABASE_BACKUP_EXECUTABLES[$driverName][$type])),
+            '{{BINARY}}' => escapeshellarg($this->getBinary(DATABASE_BACKUP_EXECUTABLES[$driverName][$OperationType->value])),
             '{{AUTH_FILE}}' => method_exists($this, 'getAuthFilePath') && $this->getAuthFilePath() ? escapeshellarg($this->getAuthFilePath()) : '',
             '{{DB_USER}}' => $this->getConfig('username'),
             '{{DB_PASSWORD}}' => $this->getConfig('password') ? ':' . $this->getConfig('password') : '',
@@ -92,7 +86,7 @@ abstract class AbstractDriver implements EventListenerInterface
             '{{DB_NAME}}' => $this->getConfig('database'),
         ];
         /** @var string $exec */
-        $exec = Configure::readOrFail('DatabaseBackup.' . $driverName . '.' . $type);
+        $exec = Configure::readOrFail('DatabaseBackup.' . $driverName . '.' . $OperationType->value);
 
         return str_replace(array_keys($replacements), $replacements, $exec);
     }
@@ -103,13 +97,15 @@ abstract class AbstractDriver implements EventListenerInterface
      * @param string $filename Filename where you want to export the database
      * @return string
      * @throws \LogicException
+     * @throws \ValueError With a filename that does not match any supported compression.
      */
     public function getExportExecutable(string $filename): string
     {
-        $exec = $this->getExecutable('export');
-        $compression = self::getCompression($filename);
-        if ($compression) {
-            $exec .= ' | ' . escapeshellarg($this->getBinary($compression));
+        $exec = $this->getExecutable(OperationType::Export);
+
+        $Compression = Compression::fromFilename($filename);
+        if ($Compression !== Compression::None) {
+            $exec .= ' | ' . escapeshellarg($this->getBinary(lcfirst($Compression->name)));
         }
 
         return $exec . ' > ' . escapeshellarg($filename);
@@ -124,12 +120,13 @@ abstract class AbstractDriver implements EventListenerInterface
      */
     public function getImportExecutable(string $filename): string
     {
-        $exec = $this->getExecutable('import');
-        $compression = self::getCompression($filename);
-        if ($compression) {
+        $exec = $this->getExecutable(OperationType::Import);
+
+        $Compression = Compression::fromFilename($filename);
+        if ($Compression !== Compression::None) {
             return sprintf(
                 '%s -dc %s | ',
-                escapeshellarg($this->getBinary($compression)),
+                escapeshellarg($this->getBinary(lcfirst($Compression->name))),
                 escapeshellarg($filename)
             ) . $exec;
         }