+++ src/OperationType.php
@@ -1,0 +1,16 @@
+<?php
+declare(strict_types=1);
+
+namespace DatabaseBackup;
+
+/**
+ * OperationType.
+ *
+ * @since 2.13.5
+ */
+enum OperationType: string
+{
+    case Export = 'export';
+
+    case Import = 'import';
+}